% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_data.R
\name{download_sp500_hist_data}
\alias{download_sp500_hist_data}
\title{Download all the S&P historical data it use the  get_one_ticker function to all the S&P500 tickers}
\usage{
download_sp500_hist_data(folder_to_save, start_date = "1900-01-01",
  end_date = Sys.Date(), mas = c(50, 100, 200), emas = c(3, 5, 7,
  14), keepma = T, keepema = T, addmacd = T, addrsi = T,
  remove_nas = T, calc_diff = T, add_boilinger = T,
  add_macd_cross = T)
}
\arguments{
\item{folder_to_save}{Path of the folder where data should be save}

\item{start_date}{The first date of the historical data}

\item{end_date}{The last date of the data}

\item{mas}{List of the simple moving averages to calculate}

\item{emas}{List of the exponentional moving averages to calculate}

\item{keepma}{Logical value if it is F it will delete the actual ma values}

\item{keepema}{Logical value if it is F it will delete the actual ema values}

\item{addmacd}{Calculate macd nFast=12, nSlow=26,nSig=9}

\item{addrsi}{Add relativ sthrength index  n=14}

\item{remove_nas}{Remove all na values after calculating every sma, ema, rsi, etc}

\item{calc_diff}{If it false it does not calculate the sma and ema}

\item{add_boilinger}{Adding the diff from the boilinger bands}

\item{add_macd_cross}{Adding if the macd is crossed}
}
\description{
Download all the S&P historical data it use the  get_one_ticker function to all the S&P500 tickers
}
