% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_report.R
\name{get_ggplot_of_ticker_to_slack}
\alias{get_ggplot_of_ticker_to_slack}
\title{Post a picture of a ticker to slack}
\usage{
get_ggplot_of_ticker_to_slack(ticker, slack_token, slack_chanel_id)
}
\arguments{
\item{ticker}{Ticker id of the company}

\item{slack_token}{Slack token}

\item{slack_chanel_id}{Slack chanel id(not the name of it)}
}
\description{
Post a picture of a ticker to slack
}
