% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_report.R
\name{get_info_plot}
\alias{get_info_plot}
\title{Get info of a ticker from tradingview}
\usage{
get_info_plot(my_ticker, show_date)
}
\arguments{
\item{show_date}{start date of the plot}

\item{Ticker}{The id of the company}

\item{my_json_string}{json string from inspect view at the bottom}
}
\description{
Get info of a ticker from tradingview
}
