% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradingview.R
\name{get_performance_table}
\alias{get_performance_table}
\title{Return all the ticker from tradingview with performance data}
\usage{
get_performance_table()
}
\description{
Return all the ticker from tradingview with performance data
}
