% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_report.R
\name{get_summary_of_ticker}
\alias{get_summary_of_ticker}
\title{Get info of a ticker from tradingview}
\usage{
get_summary_of_ticker(ticker)
}
\arguments{
\item{Ticker}{The id of the company}
}
\description{
Get info of a ticker from tradingview
}
