% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradingview.R
\name{get_tradingview_data_from_json_string}
\alias{get_tradingview_data_from_json_string}
\title{Return the sector industry data}
\usage{
get_tradingview_data_from_json_string(my_json_string)
}
\arguments{
\item{my_json_string}{json string from inspect view at the bottom}
}
\description{
Return the sector industry data
}
