% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack_report.R
\name{get_info_plot}
\alias{get_info_plot}
\title{Get info plot}
\usage{
get_info_plot(my_ticker)
}
\arguments{
\item{Ticker}{The id of the company}

\item{my_json_string}{json string from inspect view at the bottom}

\item{show_date}{start date of the plot}
}
\description{
Get info plot
}
